+++ docs/sphinx/changelog.rst
@@ -21,6 +21,7 @@ New Features
 - Added setting and checking the tolerance of the non-linear solve (:merge:`14`). By `Nathan Miller`_.
 - Added setting and checking the tolerance for the line-search of the non-linear solve (:merge:`14`). By `Nathan Miller`_.
 - Added the decomposition of the unknown vector and its application to the solution quantities (:merge:`14`). By `Nathan Miller`_.
+- Added the solution of the non-linear problem (:merge:`14`). By `Nathan Miller`_.
 
 Internal Changes
 ================